+++ assets/javascripts/kitten/components/molecules/buttons/social-button-icon-words/index.js
@@ -6,28 +6,28 @@ import { LinkedinIcon } from '../../../../components/graphics/icons/linkedin-ico
 import { InstagramIcon } from '../../../../components/graphics/icons/instagram-icon'
 
 export const FacebookButtonIconWords = ({ children, ...props }) => (
-  <Button {...props} modifier="social_facebook">
+  <Button {...props} modifier="beryllium">
     <FacebookIcon height="14" width="7" />
     <span>{children}</span>
   </Button>
 )
 
 export const TwitterButtonIconWords = ({ children, ...props }) => (
-  <Button {...props} modifier="social_twitter">
+  <Button {...props} modifier="beryllium">
     <TwitterIcon height="12" width="15" />
     <span>{children}</span>
   </Button>
 )
 
 export const InstagramButtonIconWords = ({ children, ...props }) => (
-  <Button {...props} modifier="social_instagram">
+  <Button {...props} modifier="beryllium">
     <InstagramIcon height="16" width="16" />
     <span>{children}</span>
   </Button>
 )
 
 export const LinkedinButtonIconWords = ({ children, ...props }) => (
-  <Button {...props} modifier="social_linkedin">
+  <Button {...props} modifier="beryllium">
     <LinkedinIcon height="12" width="12" />
     <span>{children}</span>
   </Button>