+++ assets/javascripts/kitten/components/molecules/cards/team-card/components/icons.js
@@ -73,7 +73,7 @@ const SocialLinks = ({ links = [] }) => (
           key={link.name}
           aria-label={link.name}
           href={link.href}
-          modifier={`social_${link.name}`}
+          modifier="hydrogen"
           target="_blank"
           rel="noopener"
           className="k-TeamCardIcons__buttonIcon"